Mass bugfixes
