+++ dalmatian/__init__.py
@@ -1,3 +1,3 @@
 __version__ = "0.0.16"
 from .core import *
-from .base import *
+from .wmanager import *
