+++ setup.py
@@ -41,7 +41,7 @@ setup(
         'google-cloud-storage>=1.13.2',
         'agutil>=4.0.2',
         'crayons>=0.2.0',
-        'hound>=0.1.9',
+        'hound>=0.2.0',
     ],
     classifiers = [
         "Programming Language :: Python :: 3",
